+++ admin/create.inc.php
@@ -112,7 +112,8 @@ if ($_POST && !process_fields($row["fields"]) && !Admin::get()->getErrors()) {
 		}
 		$name = trim($row["name"]);
 
-		queries_redirect(ME . (support("table") ? "table=" : "select=") . urlencode($name), $message, alter_table(
+		$location = ME . (support("table") ? "table=" : "select=") . urlencode($name);
+		$result = alter_table(
 			$TABLE,
 			$name,
 			(DIALECT == "sqlite" && ($use_all_fields || $foreign) ? $all_fields : $fields),
@@ -122,7 +123,13 @@ if ($_POST && !process_fields($row["fields"]) && !Admin::get()->getErrors()) {
 			($row["Collation"] && $row["Collation"] != $table_status["Collation"] ? $row["Collation"] : ""),
 			($row["Auto_increment"] != "" ? number($row["Auto_increment"]) : ""),
 			$partitioning
-		));
+		);
+
+		if ($result && !Queries::$queries) {
+			redirect($location); // Nothing was changed.
+		}
+
+		queries_redirect($location, $message, $result);
 	}
 }